fix delay time only 4ms max add timeout to wait for drdy pin

+++ lib/OS/StallardOSHelpers.h
@@ -72,7 +72,7 @@ struct function_struct
   volatile uint8_t waitingForSemaphore; //Is task waiting for a semaphore
   volatile uint16_t *semVal;
 #endif
-  volatile uint32_t continueInUS; //Delay amount
+  volatile uint64_t continueInUS; //Delay amount
 };
 
 #endif
